+++ packages/docker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @computesdk/docker
 
+## 1.2.17
+
+### Patch Changes
+
+- Updated dependencies [208a400]
+  - computesdk@1.16.0
+
 ## 1.2.16
 
 ### Patch Changes