+++ packages/fly/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/fly
 
+## 1.1.18
+
+### Patch Changes
+
+- Updated dependencies [208a400]
+  - computesdk@1.16.0
+  - @computesdk/provider@1.0.12
+
 ## 1.1.17
 
 ### Patch Changes